Move email functionality to modules
